user app loader, syscall write and exit!!!

+++ rconfig/src/lib.rs
@@ -23,30 +23,44 @@ pub mod ios {
 /* Interface of ROS and RAPPS */
 pub mod syscall {
     /// syscall number
-    pub const SYSCALL_EXIT: usize = 93;
-    pub const SYSCALL_WRITE: usize = 64;
-    pub const SYSCALL_READ: usize = 63;
-    pub const SYSCALL_OPEN: usize = 56;
-    pub const SYSCALL_CLOSE: usize = 57;
-    pub const SYSCALL_FORK: usize = 220;
-    pub const SYSCALL_EXEC: usize = 221;
-    pub const SYSCALL_WAITPID: usize = 260;
-    pub const SYSCALL_GETPID: usize = 172;
-    pub const SYSCALL_SLEEP: usize = 101;
+    pub const SYSCALL_EXIT: u32= 93;
+    pub const SYSCALL_WRITE: u32= 64;
+    pub const SYSCALL_READ: u32= 63;
+    pub const SYSCALL_OPEN: u32= 56;
+    pub const SYSCALL_CLOSE: u32= 57;
+    pub const SYSCALL_FORK: u32= 220;
+    pub const SYSCALL_EXEC: u32= 221;
+    pub const SYSCALL_WAITPID: u32= 260;
+    pub const SYSCALL_GETPID: u32= 172;
+    pub const SYSCALL_SLEEP: u32= 101;
     /// syscall register index
-    pub const SYSCALL_REG_NUM: usize = 17; // a7
-    pub const SYSCALL_REG_ARG0: usize = 10; // a0
-    pub const SYSCALL_REG_ARG1: usize = 11;
-    pub const SYSCALL_REG_ARG2: usize = 12;
-    pub const SYSCALL_REG_ARG3: usize = 13;
-    pub const SYSCALL_REG_ARG4: usize = 14;
-    pub const SYSCALL_REG_ARG5: usize = 15;
-    pub const SYSCALL_REG_ARG6: usize = 16;
-    pub const SYSCALL_REG_RET: usize = 10;
+    pub const SYSCALL_REG_NUM: u32= 17; // a7
+    pub const SYSCALL_REG_ARG0: u32= 10; // a0
+    pub const SYSCALL_REG_ARG1: u32= 11;
+    pub const SYSCALL_REG_ARG2: u32= 12;
+    pub const SYSCALL_REG_ARG3: u32= 13;
+    pub const SYSCALL_REG_ARG4: u32= 14;
+    pub const SYSCALL_REG_ARG5: u32= 15;
+    pub const SYSCALL_REG_ARG6: u32= 16;
+    pub const SYSCALL_REG_RET: u32= 10;
 }
 /* Layout of OS kernel */
-// TODO: a graph
 pub mod layout {
+    //  draw a picture to show the layout of kernel memory space
+    /*
+        * 0xffffffff +------------------+ <- 0xffffffff (4GB)
+        * 0xa0000000 |     mmio         | <- DEVICE_BASE
+        * 0x84000000 |     heap         | 
+        * 0x80000000 |     kernel       | <- KERNEL_START
+        *            |------------------| 
+        *            |                  |
+        *            |------------------| 
+        *            |       ...        |
+        *            |------------------| 
+        *            |                  |
+        * 0x00000000 +------------------+ <- 0x00000000
+     */
+
     pub const KERNEL_START: usize = 0x80000000;
     pub const KERNEL_HEAP_START: usize = 0x84000000;
     pub const KERNEL_HEAP_END: usize = 0xa0000000;
@@ -56,7 +70,7 @@ pub mod layout {
 
 /* Standard input/output/error settings */
 pub mod std_io {
-    pub const STDIN: usize = 0;
-    pub const STDOUT: usize = 1;
-    pub const STDERR: usize = 2;
+    pub const STDIN: u32= 0;
+    pub const STDOUT: u32= 1;
+    pub const STDERR: u32= 2;
 }
